+++ CryCommon/VertexFormats.h
@@ -13,6 +13,8 @@
 #ifndef VERTEXFORMATS_H
 #define VERTEXFORMATS_H
 
+#include <stddef.h>  // for offsetof
+
 #if _MSC_VER > 1000
 # pragma once
 #endif
@@ -291,17 +293,17 @@ const int g_VertFormatUVOffsets[] =
 	-1, // no UVs in this format - invalid format
 	-1, // VERTEX_FORMAT_P3F=1,                // shadow volumes (12 bytes)
 	-1, // VERTEX_FORMAT_P3F_COL4UB=2,         // usually terrain (16 bytes)
-	(int)&(((struct_VERTEX_FORMAT_P3F_TEX2F*)0)->st[0]), // VERTEX_FORMAT_P3F_TEX2F=3,          // everything else (20 bytes)
-	(int)&(((struct_VERTEX_FORMAT_P3F_COL4UB_TEX2F*)0)->st[0]), // VERTEX_FORMAT_P3F_COL4UB_TEX2F=4,   // usually plants (24 bytes)
-	(int)&(((struct_VERTEX_FORMAT_TRP3F_COL4UB_TEX2F*)0)->st[0]),// VERTEX_FORMAT_TRP3F_COL4UB_TEX2F=5, // fonts (28 bytes)
+	(int)offsetof(struct_VERTEX_FORMAT_P3F_TEX2F, st[0]), // VERTEX_FORMAT_P3F_TEX2F=3,          // everything else (20 bytes)
+	(int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB_TEX2F, st[0]), // VERTEX_FORMAT_P3F_COL4UB_TEX2F=4,   // usually plants (24 bytes)
+	(int)offsetof(struct_VERTEX_FORMAT_TRP3F_COL4UB_TEX2F, st[0]),// VERTEX_FORMAT_TRP3F_COL4UB_TEX2F=5, // fonts (28 bytes)
 	-1, // VERTEX_FORMAT_P3F_COL4UB_COL4UB=1,     
 	-1, // VERTEX_FORMAT_P3F_N=1,                
   -1, // VERTEX_FORMAT_P3F_N_COL4UB=1,                
-	(int)&(((struct_VERTEX_FORMAT_P3F_N_TEX2F*)0)->st[0]), // VERTEX_FORMAT_P3F_N_TEX2F=3,          // everything else (20 bytes)
-	(int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_TEX2F*)0)->st[0]) // VERTEX_FORMAT_P3F_N_COL4UB_TEX2F=4,   // usually plants (24 bytes)
+	(int)offsetof(struct_VERTEX_FORMAT_P3F_N_TEX2F, st[0]), // VERTEX_FORMAT_P3F_N_TEX2F=3,          // everything else (20 bytes)
+	(int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_TEX2F, st[0]) // VERTEX_FORMAT_P3F_N_COL4UB_TEX2F=4,   // usually plants (24 bytes)
   -1, // VERTEX_FORMAT_P3F_N_COL4UB_COL4UB=1,                
-  (int)&(((struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB_TEX2F*)0)->st[0]), // VERTEX_FORMAT_P3F_COL4UB_COL4UB_TEX2F=4,   // usually plants (24 bytes)
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F*)0)->st[0]), // VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F=4,   // usually plants (24 bytes)
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB_TEX2F, st[0]), // VERTEX_FORMAT_P3F_COL4UB_COL4UB_TEX2F=4,   // usually plants (24 bytes)
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F, st[0]), // VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F=4,   // usually plants (24 bytes)
 };
 
 
@@ -312,18 +314,18 @@ const int g_VertFormatRGBAOffsets[] =
 {
 	-1, // invalid format
 	-1,
-	(int)&(((struct_VERTEX_FORMAT_P3F_COL4UB*)0)->color.dcolor),
+	(int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB, color.dcolor),
 	-1,
-	(int)&(((struct_VERTEX_FORMAT_P3F_COL4UB_TEX2F*)0)->color.dcolor),
-	(int)&(((struct_VERTEX_FORMAT_TRP3F_COL4UB_TEX2F*)0)->color.dcolor),
-  (int)&(((struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB*)0)->color.dcolor),
+	(int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB_TEX2F, color.dcolor),
+	(int)offsetof(struct_VERTEX_FORMAT_TRP3F_COL4UB_TEX2F, color.dcolor),
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB, color.dcolor),
   -1,
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB*)0)->color.dcolor),
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB, color.dcolor),
   -1,
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_TEX2F*)0)->color.dcolor),
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB*)0)->color.dcolor),
-  (int)&(((struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB_TEX2F*)0)->color.dcolor),
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F*)0)->color.dcolor),
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_TEX2F, color.dcolor),
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB, color.dcolor),
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB_TEX2F, color.dcolor),
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F, color.dcolor),
 };
 
 // this is the table of offsets of normals relative to the start of the structure
@@ -338,13 +340,13 @@ const int g_VertFormatNormalOffsets[] =
 	-1, // VERTEX_FORMAT_P3F_COL4UB_TEX2F=4,
 	-1, // VERTEX_FORMAT_TRP3F_COL4UB_TEX2F=5,
   -1, // VERTEX_FORMAT_P3F_COL4UB_COL4UB,
-  (int)&(((struct_VERTEX_FORMAT_P3F_N*)0)->normal), // VERTEX_FORMAT_P3F_N=1,                
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB*)0)->normal), // VERTEX_FORMAT_P3F_N_COL4UB=1,                
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_TEX2F*)0)->normal), // VERTEX_FORMAT_P3F_N_TEX2F=3,          // everything else (20 bytes)
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_TEX2F*)0)->normal), // VERTEX_FORMAT_P3F_N_COL4UB_TEX2F=4,   // usually plants (24 bytes)
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB*)0)->normal),
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N, normal), // VERTEX_FORMAT_P3F_N=1,                
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB, normal), // VERTEX_FORMAT_P3F_N_COL4UB=1,                
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_TEX2F, normal), // VERTEX_FORMAT_P3F_N_TEX2F=3,          // everything else (20 bytes)
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_TEX2F, normal), // VERTEX_FORMAT_P3F_N_COL4UB_TEX2F=4,   // usually plants (24 bytes)
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB, normal),
   -1, 
-  (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F*)0)->normal),
+  (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F, normal),
 };
 
 static struct SBufInfoTable gBufInfoTable[] = 
@@ -353,42 +355,42 @@ static struct SBufInfoTable gBufInfoTable[] =
     0
   },
   {  //VERTEX_FORMAT_P3F
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F, x)  
     0
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_COL4UB
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_COL4UB *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB, x)  
     0,
     OOFS(color.dcolor),
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_TEX2F
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_TEX2F *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_TEX2F, x)  
     OOFS(st[0])
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_COL4UB_TEX2F
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_COL4UB_TEX2F *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB_TEX2F, x)  
     OOFS(st[0]),
     OOFS(color.dcolor)
 #undef OOFS
   },
   {  //VERTEX_FORMAT_TRP3F_COL4UB_TEX2F
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_TRP3F_COL4UB_TEX2F *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_TRP3F_COL4UB_TEX2F, x)  
     OOFS(st[0]),
     OOFS(color.dcolor),
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_COL4UB_COL4UB
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB, x)  
     0,
     OOFS(color.dcolor),
     OOFS(seccolor.dcolor),
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_N
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_N *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_N, x)  
     0,
     0,
     0,
@@ -396,7 +398,7 @@ static struct SBufInfoTable gBufInfoTable[] =
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_N_COL4UB
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB, x)  
     0,
     OOFS(color.dcolor),
     0,
@@ -404,7 +406,7 @@ static struct SBufInfoTable gBufInfoTable[] =
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_N_TEX2F
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_N_TEX2F *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_N_TEX2F, x)  
     OOFS(st[0]),
     0,
     0,
@@ -412,7 +414,7 @@ static struct SBufInfoTable gBufInfoTable[] =
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_N_COL4UB_TEX2F
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_TEX2F *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_TEX2F, x)  
     OOFS(st[0]),
     OOFS(color.dcolor),
     0,
@@ -420,7 +422,7 @@ static struct SBufInfoTable gBufInfoTable[] =
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_N_COL4UB_COL4UB
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB, x)  
     0,
     OOFS(color.dcolor),
     OOFS(seccolor.dcolor),
@@ -428,7 +430,7 @@ static struct SBufInfoTable gBufInfoTable[] =
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_COL4UB_COL4UB_TEX2F
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB_TEX2F *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_COL4UB_COL4UB_TEX2F, x)  
     OOFS(st[0]),
     OOFS(color.dcolor),
     OOFS(seccolor.dcolor),
@@ -436,7 +438,7 @@ static struct SBufInfoTable gBufInfoTable[] =
 #undef OOFS
   },
   {  //VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F
-#define OOFS(x) (int)&(((struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F *)0)->x)  
+#define OOFS(x) (int)offsetof(struct_VERTEX_FORMAT_P3F_N_COL4UB_COL4UB_TEX2F, x)  
     OOFS(st[0]),
     OOFS(color.dcolor),
     OOFS(seccolor.dcolor),